// src/app/pages/models/profile.ts
export class profile {
    status: boolean;
    message: string;
    response: any;
    transporter_id: number;
    company_name: string;
    owner_name: string;
    office_number: number;
    contact_number:number;
    country_code: number;
    email_id: string;
    transport_logo:string;
    service_type: number;
    website: string;
    pan_number:string;
    gst_number: string;
  
    }